pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ninja-payments-reconciler"
version = "0.1.0"
description = "Standalone reconciliation service for PSP polling and CRM integration"
authors = [{name = "Ninja"}]
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "httpx>=0.25.0",
    "python-dateutil>=2.8.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.20.0",
    "ruff>=0.1.0"
]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "B", "UP", "N", "C4", "SIM"]
ignore = ["E501"]

[tool.ruff.isort]
known-first-party = ["src"]
profile = "black"
